requirements.txt
python-binance==1.0.19
python-dotenv==1.0.0
aiohttp==3.9.1
websockets==12.0
pandas==2.1.4
numpy==1.26.2
anthropic==0.21.3
requests==2.31.0
python-telegram-bot==21.6
